requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
certifi==2024.2.2
charset-normalizer==3.3.2
cmake==3.28.3
contourpy==1.2.0
cycler==0.12.1
datasets==2.18.0
diffusers==0.26.3
dill==0.3.8
einops==0.7.0
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
grpcio==1.62.1
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
importlib_metadata==7.0.2
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
lit==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==5.26.0
psutil==5.9.8
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
soupsieve==2.5
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tifffile==2024.2.12
tokenizers==0.15.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2
transformers==4.38.2
triton==2.0.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.0
